refactor(fsd): extract auth and post entities
Phase C-3. Move shared domain primitives from src/domains/{auth,post}/_*
into src/entities/{auth,post}/{ui,api,model,config}/, and add public-API
index.ts barrels for both entities.

- auth: AuthHeader, AuthForm/* -> entities/auth/ui; userServices ->
  entities/auth/api; auth API types -> entities/auth/api/types;
  TOKEN_KEY -> entities/auth/config.
- post: Pagination, Editor, Form/* -> entities/post/ui; PostType ->
  entities/post/model/types; post constants -> entities/post/config.
- Imports across domains/, app/, shared/, jest config rewritten to
  @entities/* deep paths (consolidation to public-API imports deferred
  to a follow-up cleanup commit).

Verified: tsc 0 errors, lint 0 errors, 7 test suites / 44 tests pass,
next build emits all 8 routes.
