+++ src/app/(domain)/post/(list)/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Pagination from '@domains/post/_components/Pagination';
+import Pagination from '@entities/post/ui/Pagination';
 import PostListContainer from '@domains/post/list/_components/PostListContainer';
 import ListFilter from '@domains/post/list/_components/PostListFilter';
 import usePostListManager from '@domains/post/list/_hooks/usePostListManager';
