+++ src/domains/auth/signin/_components/SignInForm.tsx
@@ -1,6 +1,6 @@
-import ButtonBox from '@domains/auth/_components/AuthForm/ButtonBox';
-import EmailInput from '@domains/auth/_components/AuthForm/EmailInput';
-import PasswordInput from '@domains/auth/_components/AuthForm/PasswordInput';
+import ButtonBox from '@entities/auth/ui/AuthForm/ButtonBox';
+import EmailInput from '@entities/auth/ui/AuthForm/EmailInput';
+import PasswordInput from '@entities/auth/ui/AuthForm/PasswordInput';
 import { SignInReq } from '@domains/auth/signin/_types/api';
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
