+++ src/domains/post/create/_components/CreatePostForm.tsx
@@ -1,8 +1,8 @@
-import ButtonBox from '@domains/post/_components/Form/ButtonBox';
-import ContentsEditor from '@domains/post/_components/Form/ContentsEditor';
-import IsOpenCheckbox from '@domains/post/_components/Form/IsOpenCheckbox';
-import TitleInput from '@domains/post/_components/Form/TitleInput';
-import TypeSelect from '@domains/post/_components/Form/TypeSelect';
+import ButtonBox from '@entities/post/ui/Form/ButtonBox';
+import ContentsEditor from '@entities/post/ui/Form/ContentsEditor';
+import IsOpenCheckbox from '@entities/post/ui/Form/IsOpenCheckbox';
+import TitleInput from '@entities/post/ui/Form/TitleInput';
+import TypeSelect from '@entities/post/ui/Form/TypeSelect';
 import { CreatePostFormInput } from '@domains/post/create/_types/form';
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
