+++ src/domains/post/list/_components/PostListFilter/TypeSelect.tsx
@@ -1,4 +1,4 @@
-import { LIST_TYPE_OPTIONS } from '@domains/post/_constants/post';
+import { LIST_TYPE_OPTIONS } from '@entities/post/config/post';
 import classNames from 'classnames';
 import { useFormContext } from 'react-hook-form';
