+++ src/domains/post/update/_types/form.ts
@@ -1,4 +1,4 @@
-import { PostType } from '@domains/post/_types/postType';
+import { PostType } from '@entities/post/model/types';
 
 export interface UpdatePostFormInput {
   title: string;
