+++ src/entities/auth/config/auth.ts
