+++ src/entities/auth/ui/AuthForm/ButtonBox.tsx
