+++ src/entities/auth/ui/AuthForm/EmailInput.tsx
