+++ src/entities/auth/ui/AuthForm/NameInput.tsx
