+++ src/entities/auth/ui/AuthForm/NickNameInput.tsx
