+++ src/entities/auth/ui/AuthForm/PasswordInput.tsx
