+++ src/entities/auth/ui/AuthHeader.tsx
