+++ src/entities/post/config/post.ts
