+++ src/entities/post/index.ts
@@ -1,0 +1,16 @@
+export { default as Pagination } from './ui/Pagination';
+export { default as Editor } from './ui/Editor';
+export { default as ButtonBox } from './ui/Form/ButtonBox';
+export { default as ContentsEditor } from './ui/Form/ContentsEditor';
+export { default as IsOpenCheckbox } from './ui/Form/IsOpenCheckbox';
+export { default as TitleInput } from './ui/Form/TitleInput';
+export { default as TypeSelect } from './ui/Form/TypeSelect';
+
+export type { PostType } from './model/types';
+
+export {
+  TYPE_OPTIONS,
+  LIST_TYPE_OPTIONS,
+  PAGE_GROUP_NUMBER,
+  DUMMY_DATA,
+} from './config/post';
