+++ src/entities/post/ui/Form/IsOpenCheckbox.tsx
