+++ src/shared/lib/api.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { TOKEN_KEY } from '@domains/auth/_constants/auth';
+import { TOKEN_KEY } from '@entities/auth/config/auth';
 import { cookies } from 'next/headers';
 
 export const fetchApi = async <Response>(
